Option panel updates properly when connecting hub (midi box & firm). Midi Box greyed out when hub disconnected or paused.

+++ NeovaDashboard/Source/Common/HubConfiguration.cpp
@@ -63,7 +63,7 @@ void HubConfiguration::resetConfigWasChanged()
 
 void HubConfiguration::setMidiChannel (const uint8 newMidiChannel, bool uploadToHub)
 {
-	config.midiChannel = newMidiChannel;
+	config.midiChannel = (uint8_t) newMidiChannel;
 
 	if (uploadToHub)
 	{
